requirements.txt
binance-connector>=3.6.1,<4.0.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
python-dotenv>=1.0,<2.0
tenacity>=8.2,<9.0
pydantic>=2.6,<3.0
tabulate>=0.9,<1.0
streamlit>=1.37,<2.0
plotly>=5.20,<6.0
streamlit-autorefresh>=1.0.1,<2.0 